+++ test/complex.test.ts
@@ -1,132 +1,126 @@
-import {
-    validateComplex,
-    VALIDATE_ERROR_ARRAY_SCHEMA_MISSING,
-    VALIDATE_ERROR_ENUM_SCHEMA_MISSING,
-    VALIDATE_ERROR_OBJECT_SCHEMA_MISSING,
-    VALIDATE_ERROR_SCALAR_SCHEMA_MISSING
-} from "../src/complex";
-import { AggregationType, ComplexSchema } from "../src/schema";
+import { validateComplex } from "../src/complex";
+import { ComplexSchema } from "../src/schema";
 
 describe("Complex type definitions", () => {
     describe("A valid value", () => {
         it("should validate", () => {
-            expect(validateComplex("a", { scalarType: "string" })).toBeTruthy();
-            expect(validateComplex(0, { scalarType: "number" })).toBeTruthy();
-            expect(validateComplex(true, { scalarType: "boolean" })).toBeTruthy();
+            expect(validateComplex("a", { scalar: "string" })).toBeTruthy();
+            expect(validateComplex(0, { scalar: "number" })).toBeTruthy();
+            expect(validateComplex(true, { scalar: "boolean" })).toBeTruthy();
         });
     });
 
     describe("An invalid value", () => {
         it("should not validate", () => {
-            expect(validateComplex(0, { scalarType: "string" })).toBeFalsy();
-            expect(validateComplex(true, { scalarType: "string" })).toBeFalsy();
-            expect(validateComplex("", { scalarType: "number" })).toBeFalsy();
-            expect(validateComplex(false, { scalarType: "number" })).toBeFalsy();
-            expect(validateComplex("true", { scalarType: "boolean" })).toBeFalsy();
-            expect(validateComplex(1, { scalarType: "boolean" })).toBeFalsy();
+            expect(validateComplex(0, { scalar: "string" })).toBeFalsy();
+            expect(validateComplex(true, { scalar: "string" })).toBeFalsy();
+            expect(validateComplex("", { scalar: "number" })).toBeFalsy();
+            expect(validateComplex(false, { scalar: "number" })).toBeFalsy();
+            expect(validateComplex("true", { scalar: "boolean" })).toBeFalsy();
+            expect(validateComplex(1, { scalar: "boolean" })).toBeFalsy();
         });
     });
 
     describe("A valid array", () => {
         it("should validate", () => {
-            expect(validateComplex([], { aggregation: "array", array: { itemType: "number" } })).toBeTruthy();
+            expect(validateComplex([], { array: { itemType: "number" } })).toBeTruthy();
         });
     });
 
     describe("An invalid array", () => {
         it("should not validate", () => {
-            expect(validateComplex(["a"], { aggregation: "array", array: { itemType: "number" } })).toBeFalsy();
+            expect(validateComplex(["a"], { array: { itemType: "number" } })).toBeFalsy();
         });
     });
 
     describe("A valid object", () => {
         it("should validate", () => {
-            expect(validateComplex({}, { aggregation: "object", object: { members: {} } })).toBeTruthy();
+            expect(validateComplex({}, { object: { members: {} } })).toBeTruthy();
         });
     });
 
     describe("An invalid object", () => {
         it("should not validate", () => {
-            expect(validateComplex({}, { aggregation: "object", object: { members: { "a": "string" } } })).toBeFalsy();
+            expect(validateComplex({}, { object: { members: { "a": "string" } } })).toBeFalsy();
         });
     });
 
     describe("A valid enumerated value", () => {
         it("should validate", () => {
-            expect(validateComplex(5, { aggregation: "enum", enumOptions: [5] })).toBeTruthy();
+            expect(validateComplex(5, { enumOptions: [5] })).toBeTruthy();
         });
     });
 
     describe("An invalid enumerated value", () => {
         it("should not validate", () => {
-            expect(validateComplex(5, { aggregation: "enum", enumOptions: [] })).toBeFalsy();
+            expect(validateComplex(5, { enumOptions: [] })).toBeFalsy();
         });
     });
 
     describe("A valid string", () => {
         it("should validate", () => {
-            expect(validateComplex("", { scalarType: "string" })).toBeTruthy();
+            expect(validateComplex("", { scalar: "string" })).toBeTruthy();
         });
     });
 
     describe("An invalid string", () => {
         it("should not validate", () => {
-            expect(validateComplex(5, { scalarType: "string" })).toBeFalsy();
+            expect(validateComplex(5, { scalar: "string" })).toBeFalsy();
         });
     });
 
     describe("A valid number", () => {
         it("should validate", () => {
-            expect(validateComplex(0, { scalarType: "number" })).toBeTruthy();
+            expect(validateComplex(0, { scalar: "number" })).toBeTruthy();
         });
     });
 
     describe("An invalid number", () => {
         it("should not validate", () => {
-            expect(validateComplex("5", { scalarType: "number" })).toBeFalsy();
+            expect(validateComplex("5", { scalar: "number" })).toBeFalsy();
         });
     });
 
     describe("A valid boolean", () => {
         it("should validate", () => {
-            expect(validateComplex(true, { scalarType: "boolean" })).toBeTruthy();
+            expect(validateComplex(true, { scalar: "boolean" })).toBeTruthy();
         });
     });
 
     describe("An invalid boolean", () => {
         it("should not validate", () => {
-            expect(validateComplex(1, { scalarType: "boolean" })).toBeFalsy();
+            expect(validateComplex(1, { scalar: "boolean" })).toBeFalsy();
         });
     });
 
     describe("A valid nullable value", () => {
         it("should validate", () => {
-            expect(validateComplex(null, { scalarType: "boolean", nullable: true })).toBeTruthy();
+            expect(validateComplex(null, { scalar: "boolean", nullable: true })).toBeTruthy();
         });
     });
 
     describe("An invalid nullable value", () => {
         it("should not validate", () => {
-            expect(validateComplex(null, { scalarType: "boolean" })).toBeFalsy();
+            expect(validateComplex(null, { scalar: "boolean" })).toBeFalsy();
         });
     });
 
     describe("A valid optional value", () => {
         it("should validate", () => {
-            expect(validateComplex(undefined, { scalarType: "boolean", optional: true })).toBeTruthy();
+            expect(validateComplex(undefined, { scalar: "boolean", optional: true })).toBeTruthy();
         });
     });
 
     describe("An invalid optional value", () => {
         it("should not validate", () => {
-            expect(validateComplex(undefined, { scalarType: "boolean" })).toBeFalsy();
+            expect(validateComplex(undefined, { scalar: "boolean" })).toBeFalsy();
         });
     });
 
     describe("A valid string with a regex matcher", () => {
         it("should validate", () => {
             expect(validateComplex("I am happy", {
-                scalarType: "string",
+                scalar: "string",
                 string: {
                     matcher: /[ ]am[ ]/
                 }
@@ -137,7 +131,7 @@ describe("Complex type definitions", () => {
     describe("An invalid string with a regex matcher", () => {
         it("should not validate", () => {
             expect(validateComplex("I am happy", {
-                scalarType: "string",
+                scalar: "string",
                 string: {
                     matcher: /[ ]are[ ]/
                 }
@@ -148,7 +142,7 @@ describe("Complex type definitions", () => {
     describe("A valid string with a function matcher", () => {
         it("should validate", () => {
             expect(validateComplex("I am happy", {
-                scalarType: "string",
+                scalar: "string",
                 string: {
                     matcher: s => s.includes(" am ")
                 }
@@ -159,7 +153,7 @@ describe("Complex type definitions", () => {
     describe("An invalid string with a function matcher", () => {
         it("should not validate", () => {
             expect(validateComplex("I am happy", {
-                scalarType: "string",
+                scalar: "string",
                 string: {
                     matcher: s => s.includes(" are ")
                 }
@@ -170,7 +164,7 @@ describe("Complex type definitions", () => {
     describe("A valid string with a value matcher", () => {
         it("should validate", () => {
             expect(validateComplex("I am happy", {
-                scalarType: "string",
+                scalar: "string",
                 string: {
                     matcher: "I am happy"
                 }
@@ -181,7 +175,7 @@ describe("Complex type definitions", () => {
     describe("An invalid string with a value matcher", () => {
         it("should not validate", () => {
             expect(validateComplex("I am happy", {
-                scalarType: "string",
+                scalar: "string",
                 string: {
                     matcher: "I are happy"
                 }
@@ -192,7 +186,7 @@ describe("Complex type definitions", () => {
     describe("A valid string without a matcher", () => {
         it("should validate", () => {
             expect(validateComplex("I am happy", {
-                scalarType: "string",
+                scalar: "string",
                 string: {}
             })).toBeTruthy();
         });
@@ -201,7 +195,7 @@ describe("Complex type definitions", () => {
     describe("A valid number with a specific value", () => {
         it("should validate", () => {
             expect(validateComplex(5, {
-                scalarType: "number",
+                scalar: "number",
                 number: {
                     value: 5
                 }
@@ -212,7 +206,7 @@ describe("Complex type definitions", () => {
     describe("An invalid number with a specific value", () => {
         it("should not validate", () => {
             expect(validateComplex(4, {
-                scalarType: "number",
+                scalar: "number",
                 number: {
                     value: 5
                 }
@@ -223,7 +217,7 @@ describe("Complex type definitions", () => {
     describe("A valid number with a specific minimum", () => {
         it("should validate", () => {
             expect(validateComplex(5, {
-                scalarType: "number",
+                scalar: "number",
                 number: {
                     min: 5
                 }
@@ -234,7 +228,7 @@ describe("Complex type definitions", () => {
     describe("An invalid number with a specific minimum", () => {
         it("should not validate", () => {
             expect(validateComplex(4, {
-                scalarType: "number",
+                scalar: "number",
                 number: {
                     min: 5
                 }
@@ -245,7 +239,7 @@ describe("Complex type definitions", () => {
     describe("A valid number with a specific maximum", () => {
         it("should validate", () => {
             expect(validateComplex(5, {
-                scalarType: "number",
+                scalar: "number",
                 number: {
                     max: 5
                 }
@@ -256,7 +250,7 @@ describe("Complex type definitions", () => {
     describe("An invalid number with a specific maximum", () => {
         it("should not validate", () => {
             expect(validateComplex(6, {
-                scalarType: "number",
+                scalar: "number",
                 number: {
                     max: 5
                 }
@@ -270,14 +264,14 @@ describe("Complex type definitions", () => {
 
         it("should validate a valid value", () => {
             expect(validateComplex(MY_NUMBER, {
-                scalarType: "number",
+                scalar: "number",
                 number: { value: isMyNumber }
             })).toBeTruthy();
         });
 
         it("should not validate an invalid value", () => {
             expect(validateComplex(MY_NUMBER + 1, {
-                scalarType: "number",
+                scalar: "number",
                 number: { value: isMyNumber }
             })).toBeFalsy();
         });
@@ -285,7 +279,7 @@ describe("Complex type definitions", () => {
 
     describe("When a number's isInteger option is specified", () => {
         const schema: ComplexSchema = {
-            scalarType: "number",
+            scalar: "number",
             number: {
                 isInteger: true
             }
@@ -306,7 +300,7 @@ describe("Complex type definitions", () => {
 
     describe("When a floating point value is validated against a floating point constant", () => {
         const schema: ComplexSchema = {
-            scalarType: "number",
+            scalar: "number",
             number: {
                 value: 10.1,
                 epsilon: 1.1
@@ -326,35 +320,11 @@ describe("Complex type definitions", () => {
         });
     });
 
-    describe("A scalar type with no scalar type definition", () => {
-        it("should throw", () => {
-            expect(() => validateComplex(0, {})).toThrowError(VALIDATE_ERROR_SCALAR_SCHEMA_MISSING);
-        });
-    });
-
-    describe("An object type with no object type definition", () => {
-        it("should throw", () => {
-            expect(() => validateComplex({}, { aggregation: "object" })).toThrowError(VALIDATE_ERROR_OBJECT_SCHEMA_MISSING);
-        });
-    });
-
-    describe("An array type with no array type definition", () => {
-        it("should throw", () => {
-            expect(() => validateComplex({}, { aggregation: "array" })).toThrowError(VALIDATE_ERROR_ARRAY_SCHEMA_MISSING);
-        });
-    });
-
-    describe("An enumeration type with no enumeration type definition", () => {
-        it("should throw", () => {
-            expect(() => validateComplex({}, { aggregation: "enum" })).toThrowError(VALIDATE_ERROR_ENUM_SCHEMA_MISSING);
-        });
-    });
-
-    describe("An invalid aggregation type", () => {
+    describe("An empty schema", () => {
         it("should throw", () => {
             expect(
-                () => validateComplex([], { aggregation: "invalid" as AggregationType })
-            ).toThrowError(/Unknown aggregation type/);
+                () => validateComplex([], { })
+            ).toThrowError(/Schema only allows zero or one/);
         });
     });
 });