+++ CHANGELOG.md
@@ -4,6 +4,5 @@
 
 * Renamed TypeDef -> Schema
 * Removed deprecated settings
-* `Aggregation` is now a expressed as a string literal type
-* Deprecated `Aggregation` enumeration
+* Aggregation type is now inferred from the schema; no need to specify it directly
 * Simplified schema member naming: `arrayDef -> array, stringDef -> string` and so on